\name{graphComp-internal}

\alias{.commonEdgesInd}
\alias{.pieGlyph1}
\alias{compGraphs}
\alias{modCompare2Graphs}

\title{Internal graphComp Functions}
\description{
  Internal functions of package \pkg{graphComp}
}

\details{
  These functions are not to be called by the user.
}
\keyword{ internal }
